requirements.txt
beautifulsoup4
ffmpeg-python
requests
typer[all]
pyyaml

